Fixed format and flash scripts for Ubuntu 16.04
Due to an error of mmcblk0 created partition (not mmcblk01 and mmcblk02, but mmcblk0p1 and mmcblk0p2), I've modified the "${out}1" to "${out}p1" and ${out}2 to ${out}p2, in order to correctly detect the partition for "format_sd.sh" script completion. I've also modified the SDCARD"1" to SDCARD"p1" and SDCARD"2" to SDCARD"p2" in order to correctly detect and choose the partitions created in "format_sd.sh" script  for flashing with "flash_sd.sh" script.